drawio-image/figure directives to replace drawio
- inherits from the image/figure directives so that draw.io images
  behave the same as images/figures
- this means that all builders will be able to handle draw.io images
  without any custom code

Refer to the README for an overview of the changes compared to the old
drawio directive.

Diagrams are now also exported to the .drawio directory beside the
document. Each diagram is exported to a separate subdirectory in this directory named by the hash of the diagram filename and export
options. The base name of the exported image file is the same as that
of the draw.io file. This has these advantages:
- we don't interfere with the image/figure directive copying of image
  files to the output directory
- a diagram needs to be exported only once when two builders use the
  same export format
- more descriptive image filenames for the HTML builder