Fix issues with symlinked source files
When using sphinx-multibuild, for example, source files are symlinked. This causes issues with referencing the exported image using a relative path. Luckily, Sphinx' image/figure directives also accept an 'absolute' path which indicates the image file's path is relative to the project's source directory, so we can use that instead.